[query-service] move Worker to Worker.scala
@daniel-goldstein, you and I are both modifying this class. I want to relocate it for
ease of editing/discovery. I hope we can merge this before either of our PRs lands, so
the rebase is easier.
